Preliminary code for Rust-18-11
This code is the definition of sin itself. I wish for it to
be burned into my memory, and into the history of the internet,
for me to look back on it and cringe at its sheer awfulness.

If ye be a weary traveller from the far corners, beware of
looking into this. It's seriously absolutely stupid code.